Miscellaneous jPile clean up
* Fixed jPile integration tests

Renamed integration tests so they will be run by maven failsafe instead of maven surefire.
This is necessary because surefire tests are run in parallel, but the integration tests use the database and cannot handle being run in parallel. Also, it's useful to be able to separate out the tests that require the database from the unit tests.

Updated Table annotation on Customer to use 'customer' (instead of 'Customer').
This was causing the performance test to fail on a case-senstive MySQL database.

* Clean up jPile code

- Used Java 7 features where available (inferred generics, multi-catch)
- Resolved all Java compiler warnings
- Updated to Spring 3.2.6 to handle method deprecated in 3.1.x but not replaced until 3.2.x
- Explicitly added used dependency org.hibernate:hibernate-core
- Updated README to use Java 7 try-with-resources
- Removed unused Maven systemPropertyVariables from POM
- Fixed typos
- Loosen SingleInfileObjectLoader constructor type argument: this removes the need for SuppressWarnings("unchecked")

* FindBugs for jPile

Enable FindBugs 3.0.1 for jPile.
Running `mvn verify` or `mvn findbugs:check` will run FindBugs against the project and fail the build if any bugs are found.

Running `mvn site` will generate the FindBugs Bug Detector Report.

Fixed a unclosed ResultSet bug found by FindBugs.